02_meme_gen_starter/src/requirements.txt
lxml==4.5.0
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
six==1.14.0
